requirements.txt
# EVA Requirements
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0

# LLM Providers
groq>=0.37.0
# openai>=1.0.0        # Uncomment if using OpenAI
# anthropic>=0.18.0    # Uncomment if using Anthropic
# ollama>=0.1.0        # For local Ollama (optional)

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Semantic Splitting & Embeddings
langchain-experimental>=0.3.0   # SemanticChunker
langchain-huggingface>=0.1.0    # HuggingFace embeddings
sentence-transformers>=2.2.0    # BGE embeddings + CrossEncoder

# Document Parsing
docling>=2.0.0                  # IBM Research - accurate OCR
unstructured[pdf]>=0.16.0       # Fast document parsing

# Vector Database & Retrieval
chromadb>=0.4.0                 # Vector storage
langchain-chroma>=0.1.0         # LangChain ChromaDB integration
rank-bm25>=0.2.0                # BM25 sparse retrieval

# Math Query Processing
sympy>=1.12                     # Math expression parsing

# Web Scraping (upcoming)
# beautifulsoup4>=4.12.0
# playwright>=1.40.0
# requests>=2.31.0
